requirements.txt
Django==5.2.8
djangorestframework==3.15.2
Gunicorn==23.0.0
Pillow==10.4.0
psycopg2-binary==2.9.9
whitenoise==6.7.0
